Move TestRemappedValue.cpp to TestPatterns.cpp
