+++ test/lib/Transforms/CMakeLists.txt
@@ -7,7 +7,6 @@ add_llvm_library(MLIRTestTransforms
   TestLoopParametricTiling.cpp
   TestOpaqueLoc.cpp
   TestMemRefStrideCalculation.cpp
-  TestRemappedValue.cpp
   TestVectorizationUtils.cpp
 
   ADDITIONAL_HEADER_DIRS
